requirements.txt
GitPython==3.1.45
openai==2.7.1
tiktoken==0.12.0
regex==2025.11.3

